Fixing Roei's remarques. - functions names changed - added check for empty folder instead try/except - global var dumps list is created in the function itself (need to learn how to use it without "global" flag - changes for function descriptions and logs texts.

+++ ParserShrinkDumps_vR.py
@@ -5,56 +5,54 @@ import time
 import logging.config
 
 logging.config.fileConfig(fname=r'C:\Work\logging.config', disable_existing_loggers=False)
-
-# Get the logger specified in the config file
 logger = logging.getLogger('sampleLogger')
-
+# ToDo: turtle
 
 #  ---------------------------------------General constants-----------------------------------------------
-vvtool_output_folder = r'C:\Work\Line_P1_D1_1_0'
+work_folder = r'C:\Work\Line_P1_D1_1_0'
 pattern_dump = re.compile(r'.*\d\.dump')
 pattern_robot_data = re.compile(r'DataPos0.xml')
-dumps_list: list = []
 # --------------------------------------------Constants for parsing----------------------------------------------------
 app_path = r'C:\Work\TestingWorkbench.exe'
 raw_file = r'C:\Work\SFT_G5_HEAT_FW_CFG_Cambria.raw'
 bitmap = '0x8C8000000400001'
 
 
-#  ------------------------Search all .dumps file exclude _b.dumps -----------------------------------------------------
-def dumps_listing() -> list:
+def dumps_listing(vvtool_output_folder: str) -> list:
     """Create list of .dumps files in the output folder.
 
     Go over vvtool story output folder or all vvtool output folder. "vvtool_output_folder" constant.
-    Delete all the files and empty folders exclude:
+    Delete all the files exclude:
     - .dump files (not _b.dumps) -> add to dumps_list for the next parsing. "pattern_dump" constant.
     - DataPos0.xml (from "RobotData" folder) -> leave as is. May be it can be helpful with recordings.
+    Delete all empty (that listdir is not True (==0) folders.
     :return: list
     """
-    logger.debug('Start adding .dump files to a list and deleting all the others')
+    logger.debug(f'Start adding .dump files from {vvtool_output_folder} to a list and deleting all the others.')
+    global dumps_list
+    dumps_list = []
     for subdir, dirs, files in os.walk(vvtool_output_folder, topdown=False):
         for file in files:
             if pattern_dump.match(file):
-                logger.debug(f'Match the {pattern_dump} pattern. Added to dumps_list: {os.path.join(subdir, file)}')
+                logger.debug(f'File {file} match the {pattern_dump} pattern. '
+                             f'Added to dumps_list: {os.path.join(subdir, file)}')
                 dumps_list.append(os.path.join(subdir, file))
             elif pattern_robot_data.match(file):
-                logger.debug(f'Match the {pattern_robot_data}. Pass.')
+                logger.debug(f'File {file} match the {pattern_robot_data}. Pass.')
             else:
                 logger.debug(f'File will be deleted: {file}')
                 os.remove(os.path.join(subdir, file))
                 logger.debug(f'Deleted file: {file}')
-        try:  # ROEI - instead of try\catch I suggest to add check if the folder is empty (more readable code)
+        if not os.listdir(subdir):
             logger.debug(f'Folder will be deleted: {subdir}')
             os.rmdir(subdir)
             logger.debug(f'Deleted folder: {subdir}')
-        except OSError:
-            logger.debug(f'Can\'t delete this folder: {subdir}, may be it\'s not empty ')
-    return dumps_list  # ROEI - its global, you dont need to return it. I suggest to not use globals, create empty list
-    # in the method and return it. the next method will get it as argument.
+        else:
+            logger.debug(f'Folder: {subdir} is not empty. Will not be deleted.')
+    return dumps_list
 
 
-def testingworkbench(dump_file: str, output_folder: str, ) -> None:
-    # ROEI - I suggest to change to: runTestingworkbenchWithRecording
+def run_testingworkbench(dump_file: str, output_folder: str, ) -> None:
     """
     Run the the the testingworkbench application with needed parameters.
     Config (raw file) and extended (bitmap) parameters are predefined
@@ -74,55 +72,55 @@ def testingworkbench(dump_file: str, output_folder: str, ) -> None:
                      '-extended', bitmap])
     logger.info(f'### File is parsed ###')
 
-#  ROEI - this method need to be called: validateFwOutput
-def count_files_in_folder(dump_path) -> None:
-    # ROEI - see naming convention, you can use CamelCaseName or underline_name, but not both
-    """
-    Check the quantity and size of all (dump and parsed) files in the folder with dump. The size of a file should be
-    more than 1 kb, and the quantity of files should be
+
+def count_files_in_folder(dump_path: str) -> None:  # ROEI - this method need to be called: validateFwOutput
+    """Check the quantity and size of all (dump and parsed) files in the folder with .dump file.
+
+    The size of a file should be more than 1 kb, and the quantity of files should be
     "quantity of parsed files with the specified bitmap + 1 dump file"
-    :param dump_path:
+    :param dump_path: path to folder with .dump file and parsed .csv files
     :return: None
     """
-    output_folder = os.path.dirname(dump_path)  # ToDo: the same var exists in parsing function. How can I use only 1? #Roei - dont! it better to separate
+    logger.debug(f'Start function with parameter {dump_path}')
+    output_folder = os.path.dirname(dump_path)
     logger.debug(f'Output folder is {output_folder}')
 
     with os.scandir(output_folder) as entries:
         logger.debug(f'Folder {output_folder} contains:')
         for entry in entries:
             file_size = os.path.getsize(entry) / 1024  # ROEI - why 1024? please use constant
-            #  the second variant is os.stat(entry).st_size
             if file_size <= 1:
                 logger.warning(f'{entry.name: >{50}}, {file_size:.3f} KB. File is not parsed.')
             else:
                 logger.debug(f'{entry.name: >{50}}, {file_size:.3f} KB. ')
 
-    # ROei - isnt it the same loop as before?
+    # ROei - isn't it the same loop as before?
     path, dirs, files = next(os.walk(output_folder))
     number_files_in_folder = (len(files))
     if number_files_in_folder < 9:
         logger.warning('Missing files!')
-    #  consider doing it boolean method if the parsing failed
+    # Roie  consider doing it boolean method if the parsing failed
 
 
-def parsing() -> None:  # ROEI - I suggest to change the name to: parseShrinkRecording
-    """
-    Substitutes values from "dumps_list" files as parameters "dump_file" (-in) and "output_folder" (-out)
-    to the testingworkbench function.
-    Saves the parsed xmls in the same folder that dump file is .
+def parse_shrink_recording() -> None:
+    """Run the testingworkbench app for files from dumps_list.
+
+    Substitutes files from result of dumps_listing function as parameters "dump_file" (-in) and "output_folder" (-out)
+    to the runTestingworkbenchWithRecording function.
+    Output folder - is the same folder where the .dump file is, so all files will be in the one folder.
+    After parsing files size and quantity are checked (count_files_in_folder function)
     :return:
     """
     for dump in dumps_list:
         logger.debug(f'\n..........Parsing dump file N{dumps_list.index(dump)} from {len(dumps_list)}..........')
         output_folder = os.path.dirname(dump)
-        logger.debug(f'The output folder is {output_folder}')
-
-        testingworkbench(dump, output_folder)
+        logger.debug(f'Parsed files will be saved in {output_folder} folder.')
+        run_testingworkbench(dump, output_folder)
         count_files_in_folder(dump)
 
 
 start_time = time.time()
-dumps_listing()
-# parsing()
+dumps_listing(work_folder)
+parse_shrink_recording()
 end_time = time.time()
 logger.info(f'Script running time is {(end_time - start_time) // 60} minutes')
